DEBUG: Potential fix, match CSI2 and ov5640 formats
This is a potential fix to the broken pipe issue. This fix is a
hack, it changes the reported media bus format code for the sensor
from 2x8 to 1x16. This should allow link validation to pass, but
doesn't accurately portray how the sensor data is encoded over the
CSI2 bus since it is actually transferred one byte at a time with the
CSI2 core converting this to 1x16.

A better fix would be to write a custom link_validate function for
the CSI2 core  that allows codes that are equivalent but with different
pixel bus widths to pass validation. This should be fine because the
core converts the bus width properly when going from CSI2->AXI stream.