DEBUG: Potential fix to power issues
Currently sensor seems to not be getting initialized. This seems to
be because the ov5640 driver is expecting to have s_power get called
prior to use (perhaps by the main v4l2 bridge driver). I could not
figure out a way to easily trigger the s_power function to get called
from user space, so instead I am ending the probe function by calling
it. This is equivalent to initializing the device into a powered on state,
instead of initializing it powered off and waiting for someone else to
power it on.

Also, revert the potential fix made in commit
d8fd3548cdf47178376d61109e3d4d8fa51c418e that modified the check id
function to exit with the power regulators enabled. This will now
happen in the s_power function call.