Fix framerate set; disable init stream burst
Disable the initial stream burst that is used to coax the clock
lane into LP-11 state. This causes problems with the Xilinx
pipeline because the CSI2 RX linebuffer is filled with some data
that is never flushed. A cleaner fix will be to detect the error
this causes in the CSI2 driver, and then flush the line buffer.
Flushing the line buffer will require a GPIO reset is added to the
driver and block diagram.

Also, merge in fix to frame rate setting as described in this
mainline commit: 3b498424002755b06a86d56acf9e4c8e5e5085ff